Arreglando errores

+++ ConsultaHYIP/src/com/miguel/HYIP/Motor/Avisador.java
@@ -6,14 +6,14 @@ public class Avisador {
 
 	public static void main(String[] args) {
 		HourlyBank hourlyBank = new HourlyBank();
-		long amount = 0;
+		double amount = 0;
 		
 		if (hourlyBank.isAlive()) {
 			hourlyBank.login();
 			amount = hourlyBank.getAmount();
 			//if (amount > hourlyBank.getMinDeposit())
 			System.out.println(amount);
-			hourlyBank.logout();
+			hourlyBank.logout(); 
 			
 		}
